Rotation
Player now rotates and moves in the "right" direction.  Turning points
now have an X axis direction for the player to turn and a Z axis
direction for the player to turn.  This may change later to always turn
"left" from current direction to turn into the screen.  Depends on level
design requirements.